started day 9 assignments and fixing bugs in project

+++ Incremental Project Files/src/com/capgemini/pwa/beans/Customer.java
@@ -46,7 +46,7 @@ public class Customer {
 		this.fullName = fullName;
 	}
 
-	public int getDeposit() {
+	public double getDeposit() {
 		return deposit;
 	}
